INT-3779 & INT-4141: Flexpage get moved two pages ahead
With the same parent and moving the page
lower down on the hierarchy, the page would
actually move 1 past the target weight because
we grabbed the reference page prior to removing
the current page from the weight heirarchy.

This effectively set the moving page's weight to
+2 instead of +1 because the refpage didn't reflect
the -1 from the remove page operation.